Got first request/response to/from CHiP
The chipGetDogVersion() API is now implemented and gets back a response
that seems to make sense but I am not 100% sure it is correct since I
have nothing to check the returned versions against.

One big difference between the CHiP and MiP is that the CHiP doesn't
send back the response data as ASCII hexadecimal characters that need
to be parsed. CHiP sends back raw bytes which makes response handling
simpler.